refactor(fsm): extract verification code
extract parameter verification to a seperate function so it can be run independently of FSM
generation
